requirements.txt
# Core web framework
fastapi==0.115.7
uvicorn[standard]==0.34.0
pydantic==2.10.7
pydantic-settings==2.8.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Task queue and scheduling
celery==5.5.3
redis==6.4.0
psutil==6.1.0

# HTTP clients
httpx==0.28.1
requests==2.32.3

# Date/time handling
python-dateutil==2.9.0.post0

# Web scraping
beautifulsoup4==4.12.3
feedparser==6.0.12

# AI/ML
groq==0.13.0

# Monitoring and metrics
prometheus-client==0.21.5
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-exporter-otlp==1.30.0

# Environment and configuration
python-dotenv==1.0.1